Fix GRPOTrainer for vllm generation
Orig HF implementation for GRPOTrainer was stripping prompt of special
tokens before passing to vllm for generations. So vllm was not
outputting good completions. Created a copy of that class with bugfix.

vllm generation allows for training 1.7B models with 4096 completion
tokens. Added configs for Qwen3-0.6B and Qwen3-1.7B