Add Reasoning Gym data for training, fix GSM-Inf reward (#84)
- Add scripts to create reasoning gym training data
(family_relationships fully supported, knights_knaves data generated but
CoT examples not created). Add training configs for Qwen models.
- Major bugfix in GSM-infinite. Found that numerical ground-truth
answers of multiple digits like "12" and "123" were being modified to
"1,2" and "1,2,3" and the reward signal was incorrect for those. Use the
reward on single_answer function instead, and change configs to use
`exact_match` instead of `f1` to make the reward function explicit. `f1`
on single string would have been the same as `exact_match`, but the
latter makes it explicit with the paper.

---------

Co-authored-by: Anmol Kabra <ak2426@cornell.edu>
Co-authored-by: Albert Gong <ag2435@cornell.edu>
Co-authored-by: Anmol Kabra <22307419+anmolkabra@users.noreply.github.com>
Co-authored-by: Yilun Yin <yy958@g2-login-02.coecis.cornell.edu>
Co-authored-by: Yilun Yin <yy958@nikola-compute-14.cs.cornell.edu>